update-report-endpoints

+++ src/app/api/reports/[id]/route.ts
@@ -6,34 +6,38 @@ import { NextRequest, NextResponse } from "next/server";
 
 export async function PATCH(req: NextRequest) {
   try {
-    // const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
+    const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
 
-    // if (!token || token.role !== "dentist") {
-    //   return NextResponse.json(createResponse(false, "Unauthorized", null), {
-    //     status: 401,
-    //   });
-    // }
+    if (!token || token.role !== "dentist") {
+      return NextResponse.json(createResponse(false, "Unauthorized", null), {
+        status: 401,
+      });
+    }
+
+    const dentistId = token.sub;
 
     const reportId = req.nextUrl.pathname.split("/").pop();
+    const report = await prisma.report.findUnique({
+      where: { id: reportId },
+    });
 
     if (!reportId || !isValidCuid(reportId)) {
       return NextResponse.json(
         createResponse(false, "Invalid Report Id.", null),
-        { status: 400 }
+        { status: 400 },
       );
     }
 
-    const updatedReport = await req.json();
-
+    if (report?.dentistId !== dentistId) {
+      NextResponse.json(
+        createResponse(false, "Unauthorized to update this report.", null),
+        { status: 403 },
+      );
+    }
     await prisma.report.update({
       where: { id: reportId },
-      data: updatedReport,
+      data: { dentistId },
     });
-
-    return NextResponse.json(
-      createResponse(true, "Resport updated successfully.", null),
-      { status: 200 }
-    );
   } catch (error) {
     console.log("Error in updating report ", error);
     const errorMessage = error instanceof Error ? error.message : String(error);
@@ -47,7 +51,20 @@ export async function DELETE(req: NextRequest) {
   try {
     const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
 
-    if (!token || token.role !== "dentist") {
+    if (!token) {
+      return NextResponse.json(createResponse(false, "Unauthorized", null), {
+        status: 401,
+      });
+    }
+    if (token.role !== "dentist") {
+      return NextResponse.json(createResponse(false, "Unauthorized", null), {
+        status: 403,
+      });
+    }
+
+    const dentistId = token.sub;
+
+    if (!dentistId) {
       return NextResponse.json(createResponse(false, "Unauthorized", null), {
         status: 401,
       });
@@ -58,7 +75,7 @@ export async function DELETE(req: NextRequest) {
     if (!reportId || !isValidCuid(reportId)) {
       return NextResponse.json(
         createResponse(false, "Invalid Report Id.", null),
-        { status: 400 }
+        { status: 400 },
       );
     }
 
@@ -69,18 +86,24 @@ export async function DELETE(req: NextRequest) {
     if (!report) {
       return NextResponse.json(
         createResponse(false, "Report with this Id does not exist.", null),
-        { status: 404 }
+        { status: 404 },
       );
     }
 
-    await prisma.report.delete({
-      where: { id: reportId },
-    });
-
-    return NextResponse.json(
-      createResponse(true, "Report deleted successfully.", null),
-      { status: 200 }
-    );
+    if (report.dentistId === dentistId) {
+      await prisma.report.delete({
+        where: { id: reportId },
+      });
+      return NextResponse.json(
+        createResponse(true, "Report deleted successfully.", null),
+        { status: 200 },
+      );
+    } else {
+      return NextResponse.json(
+        createResponse(false, "Unauthorized to delete this report.", null),
+        { status: 403 },
+      );
+    }
   } catch (error) {
     console.log("Error in deleting report ", error);
     const errorMessage = error instanceof Error ? error.message : String(error);
